# docs/swagger.yaml
basePath: /
host: localhost:8080
info:
  contact:
    name: API Support
    url: http://localhost:8080/home
  description: Go Gin 框架示例项目 API 文档，包含用户认证、文章管理、标签管理等接口
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gin Demo API
  version: "1.0"
paths:
...
  /tags/article-tags:
    get:
      description: 返回指定文章关联的所有标签
      parameters:
      - description: 文章ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 标签列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的文章ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 获取文章标签
      tags:
      - 标签管理
  /tags/create:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 创建新标签。如果同名标签已存在，直接返回已有标签而不重复创建。
      parameters:
      - description: 标签名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功或返回已有标签
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建标签
      tags:
      - 标签管理
  /tags/update-article-tags:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: 全量替换指定文章的标签。传入的新标签列表会完全替换旧标签。不存在的标签会自动创建。
      parameters:
      - description: 文章ID
        in: formData
        name: id
        required: true
        type: integer
      - collectionFormat: csv
        description: 新的标签名称列表
        in: formData
        items:
          type: string
        name: tags
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: 参数错误
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 文章不存在
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: 更新文章标签
      tags:
      - 标签管理
...
securityDefinitions:
  BearerAuth:
    description: '请在请求头中携带 Token: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"